Rewrite deinflector to use a minimal perfect hash function
Use gperf to create a minimal perfect hash function for the deinflection table, which allows a deinflection in O(1) time in the number of deinflections.

In a future commit, the deinflection table needs to be expanded by a rules_in and a rules_out array, to filter out impossible deinflections.